+++ src/BadgeSelectItemBuilder/index.d.ts
@@ -1,15 +1,19 @@
 import * as React from 'react';
-import BadgeSelectOption from '../BadgeSelect/index';
+import { BadgeSkin } from '../Badge';
 
 
 export type BadgeSelectItemBuilderFn = (data: {
   id: string | number;
+  dataHook?: string;
+  className?: string;
   text: React.ReactNode;
   subtitle?: string;
   skin: BadgeSelectItemSkin;
   ellipsis?: boolean;
+  disabled?: boolean;
 }) => {
-  overrideStyle: true;
+  disabled: boolean;
+  overrideOptionStyle: true;
   id: string | number;
   value: (props?: Partial<BadgeSelectOption>) => React.ReactNode;
 };
@@ -36,3 +40,16 @@ export type BadgeSelectItemSkin =
   | 'premium';
 
 export const badgeSelectItemBuilder: BadgeSelectItemBuilderFn;
+
+export interface BadgeSelectOption {
+  id: string;
+  className?: string;
+  dataHook?: string;
+  skin: BadgeSkin;
+  selected?: boolean;
+  disabled?: boolean;
+  highlighted?: boolean;
+  text: React.ReactNode;
+  subtitle?: string;
+  ellipsis?: boolean;
+}