+++ src/Carousel/Carousel.js
@@ -2,6 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ChevronLeftLarge from 'wix-ui-icons-common/ChevronLeftLarge';
 import ChevronRightLarge from 'wix-ui-icons-common/ChevronRightLarge';
+import ChevronLeftLargeSmall from 'wix-ui-icons-common/ChevronLeftLargeSmall';
+import ChevronRightLargeSmall from 'wix-ui-icons-common/ChevronRightLargeSmall';
 // This is here and not in the test setup because we don't want consumers to need to run it as well
 import '../common/match-media-register';
 import Slider from 'react-slick';
@@ -34,28 +36,45 @@ class Carousel extends React.Component {
   static propTypes = {
     /** Applied as data-hook HTML attribute that can be used in the tests */
     dataHook: PropTypes.string,
+
     /** A single CSS class name to be appended to the Carousel's wrapper element. */
     className: PropTypes.string,
+
     /** Array of objects where each contains the `src` of an image (in \<img src="your_src" /\>) */
     images: PropTypes.array,
+
     /** Any element to render inside */
     children: PropTypes.node,
+
     /** Sets the skin of the arrow buttons */
-    buttonSkin: PropTypes.oneOf(['standard', 'inverted']),
+    buttonSkin: PropTypes.oneOf(['standard', 'inverted', 'light']),
+
     /** Images loop endlessly */
     infinite: PropTypes.bool,
+
     /** Auto-playing of images */
     autoplay: PropTypes.bool,
+
     /** Show dots */
     dots: PropTypes.bool,
+
     /** Variable width of children */
     variableWidth: PropTypes.bool,
+
     /** An index of the slide to start on */
     initialSlideIndex: PropTypes.number,
+
     /** Index change callback. `index => ...` */
     afterChange: PropTypes.func,
+
     /** Index change callback. `(oldIndex, newIndex) => ...` */
     beforeChange: PropTypes.func,
+
+    /** Sets the arrows position */
+    controlsPosition: PropTypes.oneOf(['sides', 'overlay', 'bottom', 'none']),
+
+    /** Sets the arrows position */
+    controlsSize: PropTypes.oneOf(['tiny', 'small', 'medium']),
   };
 
   static defaultProps = {
@@ -65,6 +84,8 @@ class Carousel extends React.Component {
     initialSlideIndex: 0,
     images: [],
     buttonSkin: 'standard',
+    controlsPosition: 'sides',
+    controlsSize: 'medium',
   };
 
   constructor(props) {
@@ -76,12 +97,26 @@ class Carousel extends React.Component {
   }
 
   render() {
-    const { dataHook, className, images, children } = this.props;
+    const {
+      dataHook,
+      className,
+      images,
+      children,
+      controlsPosition,
+      controlsSize,
+    } = this.props;
     const { sliderSettings } = this.state;
     const hasImages = !children && images.length > 0;
 
     return (
-      <div data-hook={dataHook} className={st(classes.root, className)}>
+      <div
+        data-hook={dataHook}
+        className={st(
+          classes.root,
+          { controlsPosition, controlsSize },
+          className,
+        )}
+      >
         <Slider {...sliderSettings}>
           {children}
           {hasImages && this._renderImages(images)}
@@ -99,6 +134,8 @@ class Carousel extends React.Component {
     initialSlideIndex,
     afterChange,
     beforeChange,
+    controlsPosition,
+    controlsSize,
   }) => {
     return {
       infinite,
@@ -116,19 +153,40 @@ class Carousel extends React.Component {
         <WrappedSliderArrow
           dataHook={dataHooks.nextButton}
           buttonSkin={buttonSkin}
-          icon={<ChevronRightLarge />}
+          arrowSize={controlsSize}
+          icon={
+            controlsSize === 'tiny' || controlsSize === 'small' ? (
+              <ChevronRightLargeSmall />
+            ) : (
+              <ChevronRightLarge />
+            )
+          }
         />
       ),
       prevArrow: (
         <WrappedSliderArrow
           dataHook={dataHooks.prevButton}
           buttonSkin={buttonSkin}
-          icon={<ChevronLeftLarge />}
+          arrowSize={controlsSize}
+          icon={
+            controlsSize === 'tiny' || controlsSize === 'small' ? (
+              <ChevronLeftLargeSmall />
+            ) : (
+              <ChevronLeftLarge />
+            )
+          }
         />
       ),
-      appendDots: pages => (
-        <Pagination className={classes.pagination} pages={pages} />
-      ),
+      arrows: controlsPosition !== 'none',
+      appendDots: pages => {
+        /*
+         * originalClassName is a workaround for stylable API extend to work and pass an extendable className.
+         * This is because react-slick overrides brutally the className prop with cloneElement().
+         */
+        return (
+          <Pagination originalClassName={classes.pagination} pages={pages} />
+        );
+      },
       customPaging: i => (
         <div
           className={classes.pageNavigation}