+++ src/Carousel/index.d.ts
@@ -1,10 +1,17 @@
 import * as React from 'react';
+import { IconButtonSkin } from '../IconButton';
+
+export type CarouselControlsPosition = 'sides' | 'overlay' | 'bottom' | 'none';
+
+export default class Carousel extends React.Component<CarouselProps> {}
+
+export type CarouselImage = Partial<HTMLImageElement> & { src: string };
 
 export interface CarouselProps {
   dataHook?: string;
   className?: string;
   images?: CarouselImage[];
-  buttonSkin?: CarouselButtonSkin;
+  buttonSkin?: IconButtonSkin;
   infinite?: boolean;
   autoplay?: boolean;
   dots?: boolean;
@@ -12,10 +19,6 @@ export interface CarouselProps {
   initialSlideIndex?: number;
   afterChange?: (currentSlide: number) => void;
   beforeChange?: (currentSlide: number, nextSlide: number) => void;
+  controlsPosition?: CarouselControlsPosition;
+  controlsSize?: 'tiny' | 'small' | 'medium';
 }
-
-export default class Carousel extends React.Component<CarouselProps> {}
-
-export type CarouselButtonSkin = 'standard' | 'inverted';
-
-export type CarouselImage = Partial<HTMLImageElement> & { src: string };